+++ components/Button/NavLink.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import { BaseButton } from './BaseButton';
 import * as React from 'react';
 
-const NavLink = ({ modifiers, ...props }: { modifiers?: any; href: string }) => {
+const NavLink = ({ modifiers, ...props }: { modifiers?: string[]; href: string }) => {
   return (
     <NavLink.Container>
       <Link {...props}>
